chore: 1 巡目レビュー指摘反映（field_size を BIGINT 化し i64 decode 型不一致を修正・非NULL round-trip テスト追加ほか）

+++ deployments/db/migrations/20260702000001_add_horse_past_runs_field_size.up.sql
@@ -1,6 +1,7 @@
 -- 近走(horse_past_runs)に出走頭数を保存する（#329 Phase1）。
 -- netkeiba 馬個別成績ページの列6(頭数 例「15」)から取得し、脚質（先行度）導出で
 -- コーナー通過順位を相対化する分母にする（1コーナー3番手は16頭立てと8頭立てで意味が違う）。
--- 既存行は NULL（backfill は近走再 fetch で埋める）。INTEGER/NULL 許容は他の任意列と同規約。
--- 冪等 ADD COLUMN IF NOT EXISTS は共有DBの先行適用列とも整合する（#331 と同方針）。
-ALTER TABLE horse_past_runs ADD COLUMN IF NOT EXISTS field_size INTEGER;
+-- 既存行は NULL（backfill は近走再 fetch で埋める）。型は当テーブルの他の整数列（round/day/
+-- distance/gate_num 等）に揃えて BIGINT にする（i64 で bind/decode する既存経路と整合。int4 だと
+-- 非NULL値の decode が i64 と型不一致になる）。冪等 ADD COLUMN IF NOT EXISTS は共有DBの先行適用列とも整合（#331 と同方針）。
+ALTER TABLE horse_past_runs ADD COLUMN IF NOT EXISTS field_size BIGINT;
